+++ packages/gatsby-plugin-fastify/src/__tests__/plugins/redirects.js
@@ -1,3 +1,5 @@
+import { StatusCodes } from "http-status-codes";
+
 describe(`Gatsby Redirects`, () => {
   it(`Should handle permanent redirect`, async () => {
     const response = await fastify.inject({
@@ -5,7 +7,7 @@ describe(`Gatsby Redirects`, () => {
       method: "GET",
     });
 
-    expect(response.statusCode).toEqual(301);
+    expect(response.statusCode).toEqual(StatusCodes.MOVED_PERMANENTLY);
     expect(response.headers.location).toEqual("/posts/page-1");
   });
 
@@ -15,7 +17,7 @@ describe(`Gatsby Redirects`, () => {
       method: "GET",
     });
 
-    expect(response.statusCode).toEqual(302);
+    expect(response.statusCode).toEqual(StatusCodes.MOVED_TEMPORARILY);
     expect(response.headers.location).toEqual("/posts/page-2");
   });
 
@@ -25,7 +27,126 @@ describe(`Gatsby Redirects`, () => {
       method: "GET",
     });
 
-    expect(response.statusCode).toEqual(307);
+    expect(response.statusCode).toEqual(StatusCodes.TEMPORARY_REDIRECT);
     expect(response.headers.location).toEqual("/posts/page-3");
+  });
+
+  it(`Should handle redirect with path params`, async () => {
+    const response = await fastify.inject({
+      url: "/redirect/a",
+      method: "GET",
+    });
+
+    expect(response.statusCode).toEqual(StatusCodes.MOVED_TEMPORARILY);
+    expect(response.headers.location).toEqual("/app/a");
+  });
+
+  it(`Should handle redirect with query string to params`, async () => {
+    const response = await fastify.inject({
+      url: "/redirect-query?letter=a",
+      method: "GET",
+    });
+
+    expect(response.statusCode).toEqual(StatusCodes.MOVED_TEMPORARILY);
+    expect(response.headers.location).toEqual("/app/a");
+  });
+
+  it(`Should handle redirect with query string to other query params`, async () => {
+    const response = await fastify.inject({
+      url: "/redirect-query-query?letter=a",
+      method: "GET",
+    });
+
+    expect(response.statusCode).toEqual(StatusCodes.MOVED_TEMPORARILY);
+    expect(response.headers.location).toEqual("/app?letter=a");
+  });
+
+  it(`Should handle redirect with catch all`, async () => {
+    const response = await fastify.inject({
+      url: "/redirect-all/a",
+      method: "GET",
+    });
+
+    expect(response.statusCode).toEqual(StatusCodes.MOVED_TEMPORARILY);
+    expect(response.headers.location).toEqual("/app/a");
+  });
+
+  it(`Should handle redirect with catch all to 1 location`, async () => {
+    const response = await fastify.inject({
+      url: "/redirect-all2/a",
+      method: "GET",
+    });
+
+    expect(response.statusCode).toEqual(StatusCodes.MOVED_TEMPORARILY);
+    expect(response.headers.location).toEqual("/app/");
+  });
+
+  it(`Should handle redirect correctly with params and catch all`, async () => {
+    const response = await fastify.inject({
+      url: "/redirect-weird/test/more-stuff",
+      method: "GET",
+    });
+
+    expect(response.statusCode).toEqual(StatusCodes.MOVED_TEMPORARILY);
+    expect(response.headers.location).toEqual("/app/test/more-stuff");
+  });
+
+  it(`Should handle redirect correctly with specific query strings`, async () => {
+    const response = await fastify.inject({
+      url: "/redirect-query-specific?id=2",
+      method: "GET",
+    });
+
+    expect(response.statusCode).toEqual(StatusCodes.MOVED_TEMPORARILY);
+    expect(response.headers.location).toEqual("/file2.pdf");
+  });
+
+  it(`Should handle colons in toPaths without splat in fromPath`, async () => {
+    const response = await fastify.inject({
+      url: "/wiki/category/url",
+      method: "GET",
+    });
+
+    expect(response.statusCode).toEqual(StatusCodes.MOVED_TEMPORARILY);
+    expect(response.headers.location).toEqual("https://en.wikipedia.org/wiki/Category:URL");
+  });
+
+  it(`Should handle splats in fromPaths with colons in toPaths`, async () => {
+    const response = await fastify.inject({
+      url: "/wiki/category/URL",
+      method: "GET",
+    });
+
+    expect(response.statusCode).toEqual(StatusCodes.MOVED_TEMPORARILY);
+    expect(response.headers.location).toEqual("https://en.wikipedia.org/wiki/Category:URL");
+  });
+
+  it(`Should handle multiple splats in fromPaths with colons in toPaths`, async () => {
+    const response = await fastify.inject({
+      url: "/wiki/Category/URL",
+      method: "GET",
+    });
+
+    expect(response.statusCode).toEqual(StatusCodes.MOVED_TEMPORARILY);
+    expect(response.headers.location).toEqual("https://en.wikipedia.org/wiki/Category:URL");
+  });
+
+  it(`Should handle colons in fromPaths that are not splats via double colon`, async () => {
+    const response = await fastify.inject({
+      url: "/Category:URL",
+      method: "GET",
+    });
+
+    expect(response.statusCode).toEqual(StatusCodes.MOVED_TEMPORARILY);
+    expect(response.headers.location).toEqual("/wiki/Category:URL");
+  });
+
+  it(`Should error when including an asterisk in toPath without a wildcard in fromPath, even if there is a splat or colon`, async () => {
+    const response = await fastify.inject({
+      url: "/some/thing/all",
+      method: "GET",
+    });
+
+    expect(response.statusCode).toEqual(StatusCodes.INTERNAL_SERVER_ERROR);
   });
 });