+++ packages/gatsby-plugin-fastify/src/utils/routes.ts
@@ -9,3 +9,17 @@ export function formatMatchPath(matchPath: string): string {
 export function removeQueryParmsFromUrl(url: string) {
   return url.split("?", 2)[0];
 }
+
+export function buildUrlFromParams(path: string, data: { [s: string]: string } = {}) {
+  return path.replace(/:(\w+)|(\*)/gi, function (match, p1, p2) {
+    if (p1 && !data[p1]) return match; // :Something in toPath does not have a splat in fromPath pass it through colon intact
+    let lookupString = p1 ?? p2;
+    let replacement = data[lookupString];
+
+    if (!replacement) {
+      throw new Error("Could not find url parameter " + lookupString + " in passed data object");
+    }
+
+    return replacement;
+  });
+}